Drow ability - conjuring an aura of darkness.
In Dungeons and Dragons, Drow have quite a few inherent abilities.
One in particular is being able to cast an aura of darkness. If playing
as a drow elf, the player can use #monster to invoke an aura of
darkness. It costs ten energy points per use, and will create an aura of
darkness similiar to that if reading a cursed scroll of light. There are
some limitations that are modelled off of the dotele() function - being
confused or stunned, being hungry, super low strength, and of course not
enough energy. As of this commit, there's no real benefit to this, but
there soon will be.

Also in this commit - starting a new game as Drow or Tortle race, the
welcome message will advise the player about using #monster particular
to that race.

Inspiration for this function comes from dNetHack.