More Drow abilities, strengths and weaknesses.
This commit basically revolves around how the Drow race operates in the
dark, or in the absence of it (light). I've also added a new trinsic
ability. Details are as follows:

* Drow take a flat -3 to-hit penalty when fighting in a lit area. But if
shrouded in darkness, they get a bonus to-hit: (u.ulevel / 3) + 2. This
equates to a +2 to-hit bonus at experience level one, and scales all the
way up to around a +15 to-hit bonus at experience level 30 (same scale
that Monks enjoy if fighting bare-handed using martial arts skill).
* Drow will heal more slowly while in a lit area. If shrouded in
darkness, the natural heal rate goes back to normal
* (Not Drow specific) some tweaks to thievery skill were made, and being
in darkness now gives the Rogue some actual benefit (moreso if that
Rogue is Drow).
* Ultravision - infravision allows the player to see a monster in the
dark, but only if that monster is infravisible. Ultravision allows
monsters to be seen in the dark no matter what. Drow inherently have
this ability, as do any drow-analogue monsters. I may confer this
ability to other monsters, need to go through my old ad&d monster
manuals and see what's appropriate
* Drow are not infravisible (that carried over from the elf template I
used in an earlier commit)

Drow race is starting to come together. Still a lot to do, but we're
getting there.