New object - potion of drow poison.
One of the things that the Drow race is notorious for are their various
potions and poisons, the most famous being their sleep poison, or 'drow
poison'. Other variants have their own methods of mimicking the Drow's
sleep-induced poison, but I don't think any of them involve the use of
an actual potion. So this is my attempt. The details:

* potion of drow poison, if ingested or hit by (potion or weapon
coated by it) will cause the target to be knocked unconscious for a
small amount of time (2-4 rounds)
* sleep resistance can negate or lessen the effect (free action does
not for this particular type of potion)
* if struck by a weapon that is 'tainted' by a potion of drow poison,
the effect is 2/3rd chance of being put to sleep for 2-4 rounds along
with 1-2 points of extra damage (non-lethal poison-based), or a 1/3rd
chance of just the small amount of extra damage
* Drow rogues will start with a potion of drow poison instead of a
potion of sickness; all drow automatically know potions of drow poison
* potions of drow poison are the rarest type of potion to spawn randomly
* Drow poison is thick like molasses, and therefore will last a bit
longer on a coated weapon vs a weapon coated with a potion of sickness

Drow poison historically has a short-lived effect, giving the one using
it just enough time to get away or to safely deliver a fatal blow. Will
be interesting to see what strategies players will devise in using this.
Expect to encounter drow monsters wielding weapons that are already
coated in drow poison. Later on, I may make it so only drow weapons can
be coated with drow poison, and/or only Drow players can utilitize it.
We'll see.

This commit also addresses a short-coming that's existed for a long
time. Early on, many weapons (non-ammo based) were given the ability to
be coated with poison. Player vs monster, this worked fine. Monster vs
player or monster vs monster, this didn't have an effect at all. That's
been addressed here.