Fix: correct Drow entry in role.c
Forgot that the capitalized bit is supposed to be the three letter short
version of the race. Also, 'droven' isn't correct, technically it's the
past participle of the word 'drive' and it hasn't been used in decades.
Using 'dark elven' is more appropriate. I'm still looking, but if anyone
knows and can reference the official term the Drow would use for 'dark
elven', please let me know.