Adjust weight of large/giant/gargantuan spiders.
Given that a large spider is about the same size as a little dog, and
giant/gargantuan spiders can be ridden, it made sense to adjust their
weight/nutrition values a bit. If you eat a giant/gargantuan spider
corpse, the amount of poison resistance given should reflect its mass.