New material type - Adamantine.
Introducing a brand new material to EvilHack - Adamantine. This metal
alloy is the material of choice for the Drow race, but Dwarves have been
known to use it as well. It is one of the hardest metal alloys known,
and is rumored to have magical abilities as well. It is extremely
difficult to create, and is therefore very rare. This commit simply
establishes the material with some basic properties, and sets the stage
for drow-specific objects to be developed. Basic stats:

* adamantine is considered a 'super material', meaning it is inherently
'fixed' and not subject to corrosion/rust
* because mithril is part of the recipe to create the adamantine alloy,
it helps make the alloy light - not as light as mithril by itself, but
significantly lighter than iron or steel
* its cost is greater than that of mithril and gold due to its rarity,
but is slightly cheaper aum for aum than platinum
* adamantine provides one level of AC protection greater than mithril,
but is less than dragonhide

This is basically the foundation for the new material. More will come in
future commits as drow objects are created and I can determine what
special properties (if any) they'll have.

One other change in this commit - the forged artifact Shadowblade is
now made of adamantine, as it's attuned to Drow.