Improvements to ultravision; fix to last commit (new material).
The initial implementation of ultravision was being able to see all
monsters in the dark. Which was great, but technically lacking, as
ultravision is supposed to let you see *everything* in the dark like it
were lit up. This commit allows for that - it covers walls, floor,
objects, traps, 'furniture' terrain, etc. Meaning a player with
ultravision will be able to see all the way down a dark corridor, or see
all of an open large space whether it's lit or not lit. Quite the
advantage actually. The effect is really cool, can't wait for players to
experience it. Big thanks to qt on figuring out how to implement this,
the display/vision code isn't easy (for me at least).

Also in this commit - a fix for the new material implementation, was
missing an entry in materialclr[].

Co-authored-by: entrez <me@entrez.cc>