Fix: temporary light/dark sources affect Drow the same as permanent sources.
The previous scenarios where being in the light or under cover of
darkness only took permanent sources into account. This commit allows
temporary sources of light/dark to have an affect on the Drow player.
This makes having a cursed magic lamp as a Drow a valuable tool,
especially if the role they play doesn't generate a lot of spell power.

Also in this commit - power regeneration for Drow only regenerates half
as fast if they're in the light.

Co-authored-by: entrez <me@entrez.cc>