Drow monsters vs player/other monsters are affected by light/dark same as a player Drow.
I also added a case I missed earlier - Drow player to-hit chances when
throwing or firing in light/dark. The bonus while in the dark isn't as
large as it is for melee, but it's still significant.

One other fix, M3_ACCURATE wasn't taken into account for monster vs
monster.