New material type - Spider Silk.
This is in preparation for Drow armor that is about to be invented.
Spider silk is very light, but also very strong, and is a valid material
for armor (non-elven) that can be made of cloth or leather. It has one
third the weight, but one extra point of AC protection vs leather. Armor
spawning randomly made of spider silk can occur (the Drow aren't as
stingy with spider silk as they are with adamantine), but it would be a
very rare occurrence. Spider silk is considered organic, and can burn.
Shopkeepers value spider silk the same as they value silver objects.