The Staff of the Archmagi will either give off light or darkness depnding on the race that wields it.
The Staff of the Archmagi is intelligent, and will attune itself to the
race of whoever wields it. Meaning it will give off light, unless
wielded by a Drow, at which point it will give off an aura of darkness.

If is currently just for player Drow, there's no handling for monster
drow at this time.