Certain Drow armor becomes brittle in the presence of light.
Traditionally, Drow armor and weapons made from adamantine cannot
tolerate being in the light (especially sunlight), and would decay over
time to the point that it no longer existed. dNetHack does this with its
drow armor (shadowsteel), I figured I'd try something a bit different.

If a piece of dark elven armor made of adamantine is in the presence of
light, it becomes brittle like glass, and will behave the same way.
Wearing drow armor made of adamantine, if it blocks an attack (player
or monster), has a 1 in 6 chance of breaking. Throwing it or kicking it,
it will break immediately.

Currently, the only drow armor objects made of adamantine are chain mail
and the bracer (shield). All other drow armor is comprised of spider
silk, and is not affected. Technically, this should affect drow weapons
made of adamantine as well (read: all of them), but I didn't go there in
this commit. It doesn't take long to break a brittle weapon when
attacking with it. I may revisit this, we'll see.