New monster: Black fungus; some fungal corpses can be used to make certain potions.
New monster - black fungus, part of the mold/fungi family, appears as a
black F. Slightly stronger than a mold, 100% passive. Anything attacking
it that isn't sleep resistant has a chance to be put to sleep for 8-10
turns. Poisonous to eat, but can confer a small amount of sleep
resistance.

The main reason this monster was created is for this purpose: by dipping
a black fungus corpse into a potion of fruit juice, you can create a
potion of drow poison. Per lore, the Drow would harvest a special type
of mushroom to make their sleep poisons. Thought it'd be neat to bring
that into the game. While I was here, I made it so dipping a gray fungus
corpse into a potion of fruit juice creates a potion of sickness.

To prevent wish farming, potions of sickness and drow poison have fixed
descriptions, neither of which are 'smoky'.