Drow monsters, driders can cast an aura of darkness around themselves.
Just like player Drow, drow monsters and driders can periodically cast
an aura of darkness centered on themselves if in the light. They can't
spam it, and quasi follows the player having to expend spell power to do
it. This gives the player a bit of time to take advantage of the
hostiles being weakened while in the light. Had to refactor litroom() to
make this happen, and there could be further uses for it down the road.