New monster: Shadow dragon.
Latest addition to the list of monsters that want to kill you - the
Shadow dragon. I *think* this is the final bit to flesh out the
inclusion of the Drow race and everything that goes with it. I may add
more, we'll see.

Shadow dragons are nasty. They have more hit dice than the average
dragon, but less than a celestial dragon. They look identical to
black dragons, and they have the typical bite/claw/engulf/breath
attacks, but their breath attack is different from anything else -
level drain (ray appears as magenta). If hit by the level drain breath
attack, you lose one level. If you're level drain resistant, you just
take physical damage. If you're not resistant but reflecting, instead of
level drain, you take damage along with some max hit point drain. This
is the same behavior for monsters being hit by a shadow dragons breath
attack, and is the first instance of reflection not being 100% for
monsters.

Shadow dragons have ultravision, and are immune to sleep/poison/level
drain. Eating their corpse confers both sleep and poison resistance (50%
per corpse). Wearing their scales as a cloak imparts ultravision and
will give off an aura of darkness. Shadow-scaled armor adds sleep and
level drain resistance. Like the drow, they are light-haters, and can be
damaged the same way drow and gremlins can.

One thing on the to-do, and I tried in this commit but it currently
doesn't work: shadow dragons are supposed to give off an aura of
darkness, just as gold dragons give off light. This has proven a hard
nut to crack, so some help may be in order, so hopefully that's
addressed soon.

Enjoy :)