Shadow dragons emit a continuous aura of darkness.
Figured it out. Shadow dragons (baby and adult) now emit an aura of
darkness, the same as how gold dragons emit light. I made some
adjustments in the range as well, a range of two was not enough.

This is kind of a scary effect. You'll see what I mean when you first
encounter a shadow dragon.