Drow and Elves, cannibalism and same-race sacrifice.
Drow and Elves are technically the same species, albeit distant cousins.
Of course when introducing a new race to the game, a separate race flag
is created for that race, and the game sees that race as not other
races. Maybe at the beginning I could have had all drow have the MH_ELF
flag and vice versa, but I'm not sure what kind of implications that
would have had for player race.

Anyways. A drow eating an elf corpse or vice versa should be counted as
cannibalism. And if that's true, then a drow sacrificing an elf corpse
and vice versa should count as same-race sacrifice. The last bit is
actually a boon for players as drow/elf and chaotic.