Fix: quest leader dying by means other than by the player directly.
Making the quest leader angry before killing the quest nemesis meant you
had to kill both to flag the quest as complete. The bug was that if a
pet or another monster killed the quest leader, that wouldn't count,
because nemdead() was checking if the player *and only the player* had
killed the quest leader. Doesn't make the game unwinnable, since the
player could potentially get another quest leader summoned via quest
artifact wishing, but it sure then makes it a giant pain in the ass.