Light haters, exploding yellow lights, and various objects that give off light.
If you or a monster is a light-hater, exploding yellow lights can cause
a small amount of damage, blind or not.

If a magic lamp (not cursed) or a powerful object that gives off light
(various dragon scales, artifacts, shield of light) is used, the light
cast will also cause light-haters a small amount of damage.