Fix: locking/unlocking containers in sokoban.
Commit b0e67c7 was a fix for locking then unlocking the Sokoban prize
closet doors, causing them to not seal up and disappear when another
closet door was opened. That fix caused a crash bug - attempting to lock
or unlock a container while in Sokoban. This commit fixes this, but I
feel like there may be a better way to do this. For now though, this
works.