Adjust feedback when freezing a liquid underfoot while riding a steed.
This is basically for when the player is wearing white dragon-scaled
armor, which allows walking over liquids and instantly freezing it as
they're walking across. The feedback given was misleading if riding a
steed, would still say 'under your feet' instead of correctly stating
under your steeds feet.

Also, if riding a clinger as a steed, the liquid below would still
freeze, that was never supposed to happen.