Crowned infidel receives sleep resistance.
Recently, an update was made so that whenever an Infidel was crowned, it
first stripped them of all of their old original racial abilities. This
needed to happen because some racial abilities (good or bad) just don't
equate as a demon. The rub is that it just wasn't abilities, but
resistances as well that were wiped. Before, this wasn't ever an issue
but it had become one now, especially for Elven and Drow infidels who
all of the sudden are no longer sleep resistant, but sleep res is
typically a thing for any other role that is crowned. Abilities, like a
centaurs speed, an elf and their search ability, or a giants
aggravation, I get that. But resistances are a whole other matter.

Shock and cold resistances are still left off here because 1) an infidel
can get shock resistance by leveling up and 2) demons and cold don't
typically get along with each other.