Fix: invoking Shadowblade to cause fear could scare monsters it shouldn't.
Resistance check wasn't included, plus monflee() doesn't check against
nearly as many monsters as onscary() does, yet both are used by various
magical items and spell to scare monsters. And on top of that, there are
places where onscary() is checked when calling monflee(), but it's not
consistent. In my opinion, if a monster is supposed to be immune to
magically induced fear, then it should be that way across the board.

Anyways. This commit addresses what was observed when invoking
Shadowblade. Casting the spell 'cause fear' or reading a scare monster
scroll now has the same checks for onscary().