Fix: vampshifted Vlad in bones.
From NetHack 3.7 git commit 17d867d (modified):

'Reported by vultur-cadens:  if Vlad made it to a level which allowed
bones and was in wolf or bat or cloud form when the hero died, he
would not be excluded from the bones file as intended.  When those
bones got loaded in another game, Vlad would become flagged as
extinct.  If that happened before his own level was reached, he
wouldn't be created when hero arrived thus there would be no way to
acquire the Candelabrum of Invocation.  (Even if his vampshifted
form had been carrying it, it wouldn't be present in the bones data
because it would have been converted into candles.)'

I added the same check for Kas and the Rat King as well - not
game-breaking for those two, but needs to be addressed regardless.