Fix: even more separate_steed_and_rider() fixes.
This stems from commit f59697e which addressed a bug with a monsters
steed not being properly zeroed out if it were killed before its rider
was. However, that caused another bug - if the steed in question were
killed and had objects in its inventory, it would cause a 'program in
disoder' as it had no place for those objects to go since the steed
carrying them was zeroed out before they could be placed. It wouldn't
cause a crash, but an error message would pop up every time the player
reloaded their game, or entered the level where it occurred.

Thanks terrapin and qt for the assist.

Co-authored-by: saltwaterterrapin <saltwaterterrapin@users.noreply.github.com>
Co-authored-by: entrez <me@entrez.cc>