Minor tweaks to receiving a spellbook via altar sacrifice.
There's already checks in place to help make sure that our hero doesn't
receive a spellbook they already have knowledge of. The biggest change
here is if our spellcasting hero has chosen to be illiterate, the
knowledge of the spell will 'fill their mind' instead of receiving a
spellbook to read.