Cleaver multi-attack changes.
Hack'EM recently pulled over code from xNetHack that prevents the
artifact Cleaver from attacking peacefuls when attacking a hostile
target and the arc swing kicks in. We discussed it a bit, and changed it
some so the not attacking peacefuls part only occurs if Cleaver is
uncursed/blessed. If cursed, it will still attack everything in its path
as before.