Greased objects have a chance to resist being stolen.
this has been unofficially on the to-do for awhile now, figured it was
about time to implement it. Various greased objects have a good chance
of resisting being stolen. Here's the breakdown:

* greased objects have a 75% chance of resisting being stolen
* if not stolen, there's a 33% chance the grease then wears off
* AD_SAMU attack type (quest artifact/invocation items/amulet of
yendor), grease will help with quest artifacts, but not with the amulet
or invocation items

This works all the way around (monster vs player, player vs monster,
monster vs monster).

This change will make cans of grease a much more valuable/desired object
that previously.