Fix: crash when Bourbon or Ozzy were level drained below zero.
Long-standing bug that was expressing itself in a way that made it
difficult to track down and fix. We were seeing coredumps where monster
aggression targets were returning NULL (if another monster killed
either of the dogs), dmonsfree errors, etc. Solution was rather simple
once the cause was discovered. Thanks terrapin for finding it.

While here - made it so neither Bourbon or Ozzy can be level drained. I
also fixed the feedback when they are defeated, as it's not always the
player directly that defeats them.

Co-authored-by: saltwaterterrapin <saltwaterterrapin@users.noreply.github.com>