Fix: weight of recipe stacks not recalculated when forging.
Discovered by terrapin - when forging and using recipes that required
multiple (stacks) of a certain object, any remaining recipe objects in
the stack weren't having their weight recalculated, resulting in an
incorrect weight for that stack.

Also in this commit - fixed an issue where dropping part of a stack
where its weight for a single object could be less than 1 aum would not
recalculate the weight of the stack when it was picked back up, causing
a heavier stack.