Fix: curses mode - prevent spurious cursor updates.
From NetHack 3.7 git commit 35273b8:

'While watching a ttyrec, I noticed strange behaviour in
the cursor updates; it was moved to approximately middle
of the map window every so often, usually when doing ranged
attacks. This wasn't really noticeable in normal gameplay,
as it was moving the cursor there, and then almost immediately
to where it was supposed to be.

I managed to trace it down to the refresh() in curses_delay_output().
That call updates the terminal to match the ncurses stdscr window,
but the stdscr window cursor position wasn't updated by NetHack.'

(cherry picked from commit 62c4b5574682750cafa49db881034d5bfd6722fd)