Dark elven bracers and dual wielding/two-handed weapons.
If we're going to properly express how bracers are worn (a pair of),
then their functionality should be addressed as well.

Bracers are worn on the forearms, and do not protude outwards like
wearing a typical shield would. Therefore, dual-wielding and
two-weaponing should be allowed while wearing a pair of them. Also
because of their nature, shield bashing shouldn't be allowed and has
been removed for bracers. Shield skill can still be trained up while
wearing them and the wearer can enjoy the extra AC benefits from the
skill.