Reduce AC bonus from shield skill if wearing bracers.
Now that bracers can be worn while twoweaponing or wielding a two-handed
weapon, it becames readily apparent that having high shield skill and
the large AC bonus that comes with it was a bit unbalanced. Made a
tweak, wearing a pair bracers still exercises shield skill, but the AC
bonus from them isn't as great as if wearing an actual shield. Still a
buff for the player depending on the role, but not one that's a over the
top.

Also in this commit, made the define 'is_bracer(otmp)', so if any other
types of bracers are made in the future, it will make implementing them
am easier process.