Move the Wizard's Tower into its own branch.
This commit takes the Wizard's Tower out of Gehennom and makes it its
own branch. It is still accessed via portal in one of the two 'fake'
wizard towers found in Gehennom. this move also reduces the total number
of levels in Gehennom by about five.

This is just the preliminary commit, as I plan on revamping the entirety
of the tower to make it more exciting and challenging experience.