Fix: Drow to-hit bonuses and penalties actually in effect, various tweaks.
Turns out when I implemented Drow race to-hit bonuses and penalties, I
only did so in the bottom line update for to-hit calculation, not in the
actual to-hit routine itself. Oops.

Also made some slight tweaks:

* Drow only receive the additional to-hit bonuses in the dark when using
drow-specific weaponry, Shadowblade, or attacking bare-handed. With any
other weapons, there is no bonus (or penalty)
* To-hit penalty for Drow while fighting in the light was increased by
one
* To-hit bonus when reaching experience level 30 was increased by one