Vampire types cannot be genocided until Vlad the Impaler is destroyed.
Thought it'd be fun to make it so vampires are immune to genocide until
the source of their creation and power were destoyed. Vampires generally
aren't much of a priority for genocide, so this isn't really a big deal.
But, this sets up destroying Vlad as an actual achievement/conduct, and
will appear as such.