Don't prevent tortles from wearing certain cursed armor.
Per discussion in IRC, it was deemed best to let the player decide if
they want to wear cursed rigid gloves/head gear or not as a tortle. If
they want to risk not being able to enshell, it's on them.

In a later commit, will most likely add a delay to enshelling if wearing
rigid armor.