Fix: certain armor would disappear on save/reload under specific conditions.
There are certain pieces of armor that are either immune to
disintegration, or are outright indestructible. However, upon
saving/reloading the game, those objects would disappear as they were
still flagged as 'in use' and marked for deletion if they had resisted
being destroyed previously. This commit fixes this.

Also in this commit - prevent the Armor of Retribution from being
destroyed from over-enchantment.