Wizard of Yendor's tower overhaul.
The Wizard of Yendor's home has undergone renovations. The old place was
a bit too restrictive, and security was somewhat lax 🙂

Quite a few changes in this commit. The Wizard's tower is now unlike
anything from its previous incarnation. Here's the breakdown:

* The Wizard's tower is still three levels as before, but each level has
been completely redesigned
* The ground level is much larger than before, with the entire base of
the tower surrounded by a moat, and the wizard has conscripted a company
of the Yendorian army as security
* The second/middle tower level is smaller than the first, and houses
various spellcasters that are there to train under the watchful eye of
the Wizard. Some monsters have been held captive here, used for
experimentation and other dark arts
* The third and final level is the very top of the tower, where the
Wizard of Yendor performs the most evil of magic spells on a magical
floating platform outside of, but connected to, the top of his tower.
This is also where the Wizard hoards his most prized treasures,
including the fabled Book of the Dead
* Entering the final level of the Wizard's tower is now counted as an
event and an achievement, and is livelogged. This also triggers a
message when entering the top level for the first time, warning the
player about falling to their death should they misstep (open air
terrain)
* How the special levels in Gehennom appear had to be tweaked a bit, as
it were possible from the initial 'wizard's tower is its own branch'
commit that the VS level could sometimes not be generated, and the
player could skip the invocation and take the stairs straight down to
the Sanctum. So now, the total number of Gehennom levels shaved off is
four instead of five, but the demon boss levels will always appear in
the correct order and with a tighter spread. The 'fake' wizard towers
are dropped from two to one, with the one guaranteed to have the portal
to the wizard tower branch, and it will always be the last Gehennom
level right before the VS level
* A fix for ants sometimes spawning in barracks was slipped into this
commit (NetHack 3.7 commit 23428d0)

The vast majority of changes are in the .des file for the tower
redesign. Storming the Wizard's tower will now be a more challenging
endeavour, but more rewarding also, and hopefully a lot more exciting
and fun than before.