Convert the named ring of invisibility found in Gollum's cave to an actual artifact.
Change the regular named ring of invisibility found in Gollum's cave to
an actual artifact, the 'One Ring'. This can't be ultra-powerful like
it's depicted in the LotR series, especially as this can be found very
early in the game. However, it's more than just a simple ring of
invisibility: it will grant stealth and searching when worn. And as an
artifact ring, it is immune to shock damage.