New branch: The Hidden Dungeon.
Introducing a new optional side branch - The Hidden Dungeon. This single
level dungeon lies directly below the castle, and was in regular use by
the castle guard until a mysterious and powerful evil force infiltrated
it. No one had ever been brave enough to reclaim the dungeon; the
entrance to the dungeon was hidden, and over time the dungeon
was forgotten, and the story passed into legend.

The Hidden Dungeon is accessible via a sealed off portal found in the
Valley of the Dead. The castle dungeon was taken over by a type of
beholder overlord, known as Tal'Gath (also sometimes referred to as
'Tal'Gath the Undying'). Ancient lore states that all of the beholders
that exist in Yendor sprang from Tal'Gath, and the only way to be rid of
all of them (genocide) is to destroy Tal'Gath.

This branch is the remains of the castle dungeon, part of it is in
fairly good shape, and the remainder has become ruin. The entire dungeon
is flooded with waste sewage from above. Various monsters inhabit this
dungeon, notably a number of beholders that protect Tal'Gath and act as
its guard.

Tal'Gath is an extremely powerful beholder, and is a force to be
reckoned with even with its gaze attacks nullified. Tal'Gath has its own
treasure stash, with a chance for a magic marker to appear in it, but
the main draw of this branch is being able to genocide beholders once
Tal'Gath has been destroyed.

This is the initial commit, some more tweaks may follow.