Finish magic chests Their addition was spread out over several commits for my sanity, but here's all the info you need to know about them: Magic chests are a new terrain type. They do not generate randomly, but show up in fixed locations (TBD by K2). They all share the same storage space, so you can put stuff in one chest and take it out of another. They display as a bright purple tool (so, usually '('), but, as a terrain type, cannot be moved. They cannot be locked, but monsters will ignore them. They cannot be destroyed and magic will not affect them in any way. Corpses will still rot and eggs will go stale inside them, and contained corpses cannot revive.
Implementation details:
Magic chests are a bit of a sleight of hand trick. There's the terrain type, which players interact with, and then a new unwishable, non-generating object called a "hidden chest." A hidden chest is created as a global variable at the beginning of the game, akin to qt's proto-patch (however, the chest itself is stored to allow an easy interface with existing looting functions since the visible manifestation is a terrain type rather than a container). It is the container that is looted whenever players interact with a magic chest square. A new where field, OBJ_SOMEWHERE, is added to accomodate it since none of the current ones fit.

Possible improvements:
-give them a unique glyph, especially in IBMgraphics
-add them to relevant maps (see "New dungeon feature: magic chests")
-playtest, playtest, playtest! I'm pretty sure I caught everything between grepping for various other where values, invent, and migrating_objs, and looking through Dynahack's implementation, but you never know.

Thanks qt and K2 for the help & ideas.