Change Convict quest artifact from an iron ball to a striped shirt.
Why? Because just about everyone ditches the Iron Ball of Liberation,
it's not practical. Changed it to be a striped shirt, has all of the
same abilities as before, but of course has to be worn vs just carried.
I have one or two more ideas for this to make it interesting, may see
those in another commit here soon.

Enjoy.