Fix: close a loophole being able to produce tame non-tameable monsters.
In tamedog() there are several monsters that are flagged as off limits
to being tamed (feeding, magical taming, etc). But creating a blessed
figurine of the same, and the player could get around that limitation.
Finally decided to dig into why that was - the restrictions in tamedog()
did not fully align with those in make_familiar(). This commit closes
that loophole. You could still get a tame 'non-tameable' by getting
lucky with a poly trap/wand/spell however, we'll leave that in place as
it's purely by chance. While fixing this, I refactored what was deemed
as non-tameable and made a define for it in mondata.h

Also in this commit, a bit of code formatting in trap.c and wield.c