Sword of Annihilation recipe changes.
Now that the SoA has been around for a bit, and has been heavily
play-tested, some realizations have come about; it's very powerful, and
is regarded as the most powerful weapon in the game (which is fine), and
it's also agreed that it's fairly easy to obtain, as both Brands are not
difficult to acquire. A change was warranted. It can stay powerful, but
the cost/risk needed to obtain it needed to be adjusted upwards.

The recipe needed now to forge the SoA is Angelslayer, and either of the
artifact body parts the Vecna dropped when destroyed. Angelslayer is not
off-limits to any roles alignment, and the Vecna artifacts are
guaranteed IF the player is willing to take the risk in obtaining them.
This prevents acquiring the SoA too early and making the early/mid game
an easy walkthrough, and also makes obtaining part of the recipe items
more like a quest vs a sacrifice farming operation.

Decided to add a chance for a forge spawning in the final Vecna level,
seemed appropriate.