New forging artifact recipe - Dichotomy.
This is what I wanted to do initially - have the Brands (fire/frost)
forged together to make a blade that did both types of damage.
SpliceHack does this with the artifact 'Frostburn' and that is
definitely the inspiration here, but I decided to do this another way,
and in doing so, address some of the shortcomings in the original
implementation.

Dichotomy is a crystal (gemstone) runesword that is created by forging
Fire Brand and Frost Brand together in a forge. The broadsword inherits
the elemental properties of both, hence the name. Those hit by this
sword will take either fire damage, cold damage, or both depending on
their resistances. If the target has neither resistance to fire/cold,
extra damage is inflicted. Objects in open inventory of the target are
subject to both fire/cold damage. Whoever wields Dichotomy is granted
both fire/cold resistance, and any objects in open inventory are
afforded the same protection.

If special damage bonus applies (target is not fully resistant to
both fire and cold attacks), there's a chance that Dichotomy will
experience a surge in elemental power, and explode a ball of fire or
cold on the target, doing extra area of effect damage.

While working on this, we noticed and fixes a few bugs: exploding types
mismatches, Grimtooth didn't have SPFX_ATTK flag like it should, several
cases where a monster using an artifact (and in one case, dragon-scaled
armor) wasn't protecting it properly, and a missing database entry.

Chose runesword (broadsword) over long sword since we already have
plenty of long sword artifact types. And every role that can gain skill
in long sword can also gain skill in broadsword. Hope this is a fun
artifact to use.