Fix: Drow shouldn't be able to handle artifacts that give off light.
It did not make sense to me that a Drow, who hates light and is even
harmed by it when in the presence of it, could handle an artifact weapon
like Sunsword or the Hammer of the Gods. This commit fixes this, and
disallows drow from wielding either weapon.

Also fixed the formatting in this bit of code, artifacts are
now expressed properly. I also added an override here for testing
purposes in wizmode.